requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
cryptography==42.0.0
httpx==0.26.0
python-multipart==0.0.6
aiosqlite==0.19.0
python-dotenv==1.0.0
gunicorn==21.2.0
bcrypt==4.1.2
pytz==2025.2